requirements.txt
python-telegram-bot==20.7
pdfrw==0.4
python-dateutil==2.8.2
pytz==2023.3